Support vector.mask + vector.multi_reduce.
Currently, vector.create_mask created special
distributed format that was suited for `transfer_read`
and `transfer_write`s.

This commit also adds a utility convert that to
conventional distributed format. We might want
to have a utility to do the other way round and
make vector.create_mask distributed generate the convetional
format. Then use that util to get it to whats
needed in transfer_read and writes.